jac_parser updated for jaclang upgrading
